% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Clean Global Environment, Plots and Console}
\usage{
clear()
}
\description{
\code{clear()} cleans three R workspaces: Global Environment, Plots and
R Console.
}
\details{
\code{clear()} removes all objects from Global Environment, clear plots
and R Console (the same as pressing \code{Ctrl + L}). It does not require
any arguments.

This can be used as an initial command line to clear the workspace as
follows:

\preformatted{
## Clear the workspace
clear()
}

\code{clear} does not interrupt logging process of \code{ilog}.
}
\examples{
# create x by generating 100 random values
x <- rnorm(100)

# plotting x
plot(x)

# calling clear to clean workspace
clear()
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{all,}
\keyword{clean,}
\keyword{clear,}
\keyword{console}
\keyword{environment,}
\keyword{global}
\keyword{plots,}
\keyword{remove}
