% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilog.R
\name{ilog}
\alias{ilog}
\alias{ilog.close}
\alias{ilogtxt}
\alias{ilogClearPrompt}
\title{Echo a copy of console to log file}
\usage{
ilog(logfile = "LOG.txt", append = FALSE)

ilog.close()

ilogtxt(cmd, res, s, vis)

ilogClearPrompt()
}
\arguments{
\item{logfile}{Name of desired log file in \code{.txt} format}

\item{append}{a logical value}

\item{cmd}{command line}

\item{res}{result}

\item{s}{others}

\item{vis}{others}
}
\description{
\code{ilog()} echo a copy of console output to log file.
}
\details{
\code{ilog} is a two-step function that allows you a record of your console.
A log is a file containing what you type and console output. If a name is not
specified, then \code{ilog} will use the name \code{<unnamed>.txt}.

\code{ilog} opens a log file and \code{ilog.close} close the file.

\strong{Warnings}:

However, clearing objects from the workspace along with hidden objects
removes \code{ilog}'s \code{.logenv} environment, hence throwing an error
when it's attemptted to be closed. An error message
\code{Error in (function (cmd, res, s, vis)  : object '.logenv' not found}
will be thrown.

In that case, console prompt is stuck at \code{log> }. If
this error occurs, use \code{ilogClearPrompt()} function to revert back to
normal.
}
\examples{
\dontrun{
## my first log
ilog("myfirstlog.txt")
str(infert)
str(iris)
ilog.close()

# in case of error: ".logenv" not found
ilogClearPrompt()
}

}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{console}
\keyword{console,}
\keyword{copy}
\keyword{echo,}
\keyword{log,}
\keyword{ouput,}
\keyword{save}
