% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printText.R
\name{printText}
\alias{printText}
\alias{printLines}
\alias{printMsg}
\alias{printWarning}
\alias{wrapText}
\alias{generateLinesDF}
\title{\code{mStats}'s Printing Functions}
\usage{
printText(x, txt, split = NULL, printDF = FALSE)

printLines(x = "_", width = 80)

printMsg(txt = NULL)

printWarning(txt = NULL)

wrapText(txt, width = 70, sep = "\\n")

generateLinesDF(x, sep = "-")
}
\arguments{
\item{x}{vector, matrix, dataframe or separator (in case of printLines)}

\item{txt}{texts}

\item{split}{separator for printText}

\item{printDF}{If yes, print as Data.frame}

\item{width}{desired character length to display}

\item{sep}{separator for line break}
}
\description{
\code{printText()}, \code{printLines()}, \code{printMsg()},
\code{printWarning()}, \code{wrapText()}

Printing Functions to format and display outputs from mStats package
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{print}
\keyword{support,}
